fetch/metadata/generated/script-element-sec-fetch-site.https.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Site - redirection from HTTPS</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(path, attributes) {
    const script = document.createElement('script');
    script.setAttribute('src', path);

    for (const [ name, value ] of Object.entries(attributes)) {
      script.setAttribute(name, value);
    }

    return new Promise((resolve, reject) => {
        script.onload = resolve;
        script.onerror = () => reject('Failed to load script');
        document.body.appendChild(script);
      })
      .then(() => script.remove());
  }

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return induceRequest(
        'https://{{host}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{hosts[alt][www]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{host}}:{{ports[https][0]}}' + path + '&mime=application/javascript',
        {}
      )
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect - classic script');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return induceRequest(
        'https://{{host}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{hosts[alt][www]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{host}}:{{ports[https][0]}}' + path + '&mime=application/javascript',
        {"type": "module"}
      )
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect - module script');
  </script>
</html>
